Box2D integration: upgrade Box2D dependency to 2.4.1 (#198)

+++ envpool/box2d/utils.cc
@@ -20,7 +20,7 @@ b2Vec2 Vec2(double x, double y) {
   return b2Vec2(static_cast<float>(x), static_cast<float>(y));
 }
 
-double Sign(double val, double eps) {
+float Sign(double val, double eps) {
   if (val > eps) {
     return 1;
   }
