+++ envpool/box2d/utils.h
@@ -17,7 +17,7 @@
 #ifndef ENVPOOL_BOX2D_UTILS_H_
 #define ENVPOOL_BOX2D_UTILS_H_
 
-#include <Box2D/Box2D.h>
+#include <box2d/box2d.h>
 
 #include <random>
 
@@ -29,7 +29,7 @@ using RandUniform = std::uniform_real_distribution<>;
 // this function is to pass clang-tidy conversion check
 b2Vec2 Vec2(double x, double y);
 
-double Sign(double val, double eps = 1e-8);
+float Sign(double val, double eps = 1e-8);
 
 }  // namespace box2d
